test: fix pixel-mouse harness assertions
Client-shell mapping uses real host chrome geometry. Headless pane
input returns false when it does not request a re-render; assert the
encoded PTY bytes instead.

+++ src/pane/input.rs
@@ -83,10 +83,10 @@ pub(super) fn promote_sgr_pixel_position(
     if width_px == 0 || height_px == 0 || cols == 0 || rows == 0 {
         return position;
     }
-    let Some(x) = crate::input::mouse::cell_origin_pixel(column, cols as u16, width_px) else {
+    let Some(x) = crate::input::mouse::cell_origin_pixel(column, cols, width_px) else {
         return position;
     };
-    let Some(y) = crate::input::mouse::cell_origin_pixel(row, rows as u16, height_px) else {
+    let Some(y) = crate::input::mouse::cell_origin_pixel(row, rows, height_px) else {
         return position;
     };
     crate::input::mouse::Position::Pixels { x, y }
